Trento.Clusters.ValueObjects.CheckResult (Trento Web v3.1.2)

Check result value object

Types

t()

@type t() :: %Trento.Clusters.ValueObjects.CheckResult{
  check_id: term(),
  result: term()
}

Functions

cast_and_validate_required_embed(changeset, field, required_fields)

cast_and_validate_required_polymorphic_embed(changeset, field, required_fields)

changeset(struct, params)

Casts the fields by using Ecto reflection, validates the required ones and returns a changeset.

new(structs)

@spec new(map() | [map()]) :: {:ok, t() | [t()]} | {:error, any()}

Returns an ok tuple if the params are valid, otherwise returns {:error, {:validation, errors}}. Accepts a map or a list of maps.

new!(params)

@spec new!(map() | [map()]) :: t() | [t()]

Returns new struct(s) if the params are valid, otherwise raises a RuntimeError.

validate_required_fields(changeset, required_fields)