Trento.Discovery (Trento Web v3.1.2)

Discovery integration context.

Types

command()

@type command() :: struct()

Functions

get_current_discovery_events()

@spec get_current_discovery_events() :: [Trento.Discovery.DiscoveryEvent.t()]

Get the discovery events that were handled to build the current state of the system.

get_discarded_discovery_events(event_number)

@spec get_discarded_discovery_events(number()) :: [
  Trento.Discovery.DiscardedDiscoveryEvent.t()
]

Get the discovery events that were dead-lettered.

handle(event)

@spec handle(map()) :: :ok | {:error, any()}

Transform a discovery in a list of commands event by using the appropriate policy. Store the event in the discovery events log for auditing purposes and dispatch the commands.

prune_discovery_events(days)

@spec prune_discovery_events(number()) :: :ok

Prune all discovery events (regular and discarded) older than the given number of days.

request_cluster_discovery(cluster_id)

@spec request_cluster_discovery(String.t()) :: :ok | {:error, any()}

Request cluster discovery

request_cluster_hosts_discovery(cluster_id)

@spec request_cluster_hosts_discovery(String.t()) :: :ok | {:error, any()}

Request cluster hosts discovery

request_saptune_discovery(host_id)

@spec request_saptune_discovery(String.t()) :: :ok | {:error, any()}

Request saptune discovery in host