Trento.Domain.Events.ClusterRolledUp (Trento Web v1.2.0)

This event is emitted when a cluster is rolled up and its stream is archived. It contains the snapshot of the cluster aggregate that will be used to restore the aggregate state.

Types

t()

@type t() :: %Trento.Domain.Events.ClusterRolledUp{
  cluster_id: term(),
  snapshot: term(),
  version: term()
}

Functions

cast_and_validate_required_embed(changeset, field, required_fields)

changeset(struct, params)

Casts the fields by using Ecto reflection, validates the required ones and returns a changeset.

new(structs)

@spec new(map() | [map()]) :: {:ok, t() | [t()]} | {:error, any()}

Returns an ok tuple if the params are valid, otherwise returns {:error, errors}. Accepts a map or a list of maps.

new!(params)

@spec new!(map() | [map()]) :: t() | [t()]

Returns new struct(s) if the params are valid, otherwise raises a RuntimeError.

upcast(params, metadata)

upcast(params, _, int)

validate_required_fields(changeset, required_fields)