Trento.Factory (Trento Web v3.1.1)

A simple Factory helper module to be used within tests to generate test data

Functions

ability_factory()

activity_log_entry_factory()

activity_log_retention_time_factory()

activity_log_settings_factory()

affected_package_factory()

affected_system_factory()

ai_configuration_creation_params_factory(attrs)

ai_user_configuration_factory(attrs)

alerting_settings_factory()

api_key_settings_factory()

application_instance_deregistered_event_factory()

application_instance_factory()

application_instance_marked_absent_event_factory()

application_instance_moved_event_factory()

application_instance_registered_event_factory()

ascs_ers_cluster_details_factory()

ascs_ers_cluster_node_factory()

ascs_ers_cluster_sap_system_factory()

bugzilla_fix_factory()

build(factory_name, attrs \\ %{})

build_list(number_of_records, factory_name, attrs \\ %{})

build_pair(factory_name, attrs \\ %{})

check_customization_applied_v1_factory()

check_customization_reset_v1_factory()

cib_resource_factory()

cluster_checks_health_changed_event_factory()

cluster_checks_selected_event_factory()

cluster_deregistered_event_factory()

cluster_details_updated_event_factory()

cluster_discovered_health_changed_event_factory()

cluster_enrichment_data_factory()

cluster_factory()

cluster_health_changed_event_factory()

cluster_host_status_changed_event_factory()

cluster_registered_event_factory()

cluster_resource_factory()

cluster_resource_parent_factory()

cluster_tombstoned_event_factory()

clustered_sap_instance_factory()

create(_)

@spec create(any()) :: no_return()

create(_, _)

@spec create(any(), any()) :: no_return()

create_list(_, _, _)

@spec create_list(any(), any(), any()) :: no_return()

create_pair(_, _)

@spec create_pair(any(), any()) :: no_return()

crm_resource_factory()

crm_resource_node_factory()

cve_factory(attrs)

database_deregistered_event_factory()

database_factory()

database_health_changed_event_factory()

database_instance_deregistered_event_factory()

database_instance_factory()

database_instance_marked_absent_event_factory()

database_instance_registered_event_factory()

database_registered_event_factory()

database_restored_event_factory()

database_tenants_updated_event_factory()

database_tombstoned_event_factory()

deregister_application_instance_command_factory()

deregister_database_instance_command_factory()

discarded_discovery_event_factory()

discovery_event_factory()

errata_details_factory()

factory(factory_name)

@spec factory(any()) :: no_return()

Raises a helpful error if no factory is defined.

failed_software_updates_discovery_result_factory()

hana_cluster_details_factory()

hana_cluster_node_factory()

heartbeat_factory()

heartbeat_failed_factory()

heartbeat_succeded_factory()

host_added_to_cluster_event_factory()

host_checks_health_changed_factory()

host_checks_selected_event_factory()

host_details_updated_event_factory()

host_discovery_event_factory()

host_factory()

host_health_changed_event_factory()

host_prometheus_targets_factory()

host_registered_event_factory()

host_removed_from_cluster_event_factory()

host_saptune_health_changed_event_factory()

host_systemd_unit_factory()

host_tombstoned_event_factory()

insert(already_built_record)

insert(already_built_record, function_opts)

insert(factory_name, attrs, opts)

insert_list(number_of_records, factory_name, attrs \\ %{})

insert_list(number_of_records, factory_name, attrs, opts)

insert_pair(factory_name, attrs \\ %{})

insert_pair(factory_name, attrs, opts)

insert_software_updates_settings(attrs \\ [])

installation_settings_factory()

operation_completed_v1_factory()

operation_completed_with_errors_v1_factory(attrs)

operation_completed_with_failed_request_v1_factory(attrs)

params_for(factory_name, attrs \\ %{})

params_with_assocs(factory_name, attrs \\ %{})

patch_for_package_factory()

personal_access_token_factory(attrs)

random_ai_model_factory(attrs)

random_ai_provider_factory(_)

register_application_instance_command_factory()

register_database_instance_command_factory()

register_host_command_factory()

register_offline_cluster_host_factory()

register_online_cluster_host_factory(attrs)

relevant_patch_factory()

rollup_sap_system_command_factory()

sap_system_deregistered_event_factory()

sap_system_factory()

sap_system_health_changed_event_factory()

sap_system_instance_factory()

sap_system_registered_event_factory()

sap_system_restored_event_factory()

sap_system_tombstoned_event_factory()

sapcontrol_process_factory()

saptune_solution_factory()

saptune_staging_factory()

saptune_status_factory()

saptune_status_updated_event_factory()

sbd_device_factory(attrs)

self_signed_certificate_factory(attrs)

sles_subscription_factory()

software_updates_discovery_cleared_event_factory()

software_updates_discovery_health_changed_event_factory()

software_updates_discovery_requested_event_factory()

software_updates_discovery_result_factory()

software_updates_settings_factory(attrs)

sso_certificates_settings_factory()

string_params_for(factory_name, attrs \\ %{})

string_params_with_assocs(factory_name, attrs \\ %{})

subscriptions_updated_event_factory()

tag_factory()

tenant_factory()

upgradable_package_factory()

user_factory()

user_identity_factory()

user_with_abilities_factory()

users_abilities_factory()