Trento.Hosts (Trento Web v3.1.2)

Provides a set of functions to interact with hosts.

Functions

by_host_id(host_id)

by_id(id)

@spec by_id(String.t()) ::
  {:ok, Trento.Hosts.Projections.HostReadModel.t()} | {:error, :not_found}

deregister_host(host_id, date_service \\ DateService)

@spec deregister_host(Ecto.UUID.t(), Trento.Support.DateService) ::
  :ok | {:error, :host_alive} | {:error, :host_not_registered}

get_all_hosts(opts \\ [])

@spec get_all_hosts(keyword()) :: [Trento.Hosts.Projections.HostReadModel.t()]

get_all_sles_subscriptions()

@spec get_all_sles_subscriptions() :: non_neg_integer()

get_host_by_id(id)

@spec get_host_by_id(Ecto.UUID.t()) ::
  Trento.Hosts.Projections.HostReadModel.t() | nil

get_hosts_for_prometheus_targets()

@spec get_hosts_for_prometheus_targets() :: [
  Trento.Hosts.Projections.HostReadModel.t()
]

Returns all hosts that are configured for Prometheus pull mode. Hosts with prometheus_mode set to :push are excluded as they push metrics directly.

request_checks_execution(host_id)

@spec request_checks_execution(String.t()) :: :ok | {:error, any()}

request_hosts_checks_execution()

request_operation(operation, host_id, params)

@spec request_operation(atom(), String.t(), map()) ::
  {:ok, String.t()} | {:error, any()}

select_checks(host_id, checks)

@spec select_checks(String.t(), [String.t()]) :: :ok | {:error, any()}