Trento.Infrastructure.Commanded.EventHandlers.StreamRollUpEventHandler (Trento Web v2.4.0)

This module is responsible for triggering the roll-up of an aggregate.

Once the stream version of an aggregate reaches a certain threshold, the roll-up process is triggered. The roll-up process consists of archiving the stream and creating a snapshot event of the aggregate, during the roll-up process, the aggregate is locked to prevent any other event from being applied.

Not all events trigger a roll-up, for instance roll-up related events are ignored to avoid side-effects.

Functions

child_spec(opts)

Provides a child specification to allow the event handler to be easily supervised.

Supports the same options as start_link/3.

The default options supported by GenServer.start_link/3 are also supported, including the :hibernate_after option which allows the process to go into hibernation after a period of inactivity.

Example

Supervisor.start_link([
  {ExampleHandler, []}
], strategy: :one_for_one)

start_link(opts \\ [])

Start an event handler GenServer process linked to the current process.

Options

:application - the Commanded application.

:name - name of the event handler used to determine its unique event store subscription.

:concurrency - determines how many processes are started to concurrently process events. The default is one process.

:consistency - one of either :eventual (default) or :strong.

:start_from - where to start the event store subscription from when first created (default: :origin).

:subscribe_to - which stream to subscribe to can be either :all to subscribe to all events or a named stream (default: :all).

The default options supported by GenServer.start_link/3 are supported, including the :hibernate_after option which allows the process to go into hibernation after a period of inactivity.