Trento.UserIdentities (Trento Web v3.1.2)

The UserIdentities context, serves as custom context for PowAssent

Functions

all(user)

Callback implementation for PowAssent.Ecto.UserIdentities.Context.all/1.

create(user, user_identity_params)

This function is deprecated. Please use `upsert/2` instead.

Callback implementation for PowAssent.Ecto.UserIdentities.Context.create/2.

create_user(user_identity_params, user_params, user_id_params)

redefining the PowAssent create user method, this is called when the user login through idp and a user identity does not exists on our database.

If a user with the same username exists on our database, the user will be recovered and associated with the idp identity, otherwise the user will be created.

delete(user, provider)

Callback implementation for PowAssent.Ecto.UserIdentities.Context.delete/2.

get_user_by_provider_uid(provider, uid)

Callback implementation for PowAssent.Ecto.UserIdentities.Context.get_user_by_provider_uid/2.

pow_assent_all(user)

pow_assent_create(user, user_identity_params)

This function is deprecated. Please use `pow_assent_upsert/2` instead.

See Trento.UserIdentities.pow_assent_upsert/2.

pow_assent_create_user(user_identity_params, user_params, user_id_params)

pow_assent_delete(user, provider)

pow_assent_get_user_by_provider_uid(provider, uid)

pow_assent_upsert(user, user_identity_params)

upsert(user, user_identity_params)

redefining the PowAssent upsert method, if a IDP user is associated with a locked user, this is called when the user login with IDP and exist in our database with or without a user identity