Trento.Users (Trento Web v3.1.1)

The Users context.

Functions

authenticate(params)

Callback implementation for Pow.Context.authenticate/1.

by_id(id)

@spec by_id(id :: non_neg_integer()) ::
  {:ok, Trento.Users.User.t()} | {:error, :not_found}

confirm_totp_enrollment(user, totp_code)

create(params)

Callback implementation for Pow.Context.create/1.

create_user(attrs)

delete(user)

Callback implementation for Pow.Context.delete/1.

delete_user(user)

get_by(clauses)

get_by function overrides the one defined in Pow.Ecto.Context, we retrieve the user by username as traditional Pow flow but we also exclude deleted and locked users

get_user(id)

initiate_totp_enrollment(user)

list_users()

maybe_disable_totp(attrs)

pow_authenticate(params)

pow_create(params)

pow_delete(user)

pow_get_by(clauses)

pow_update(user, params)

reset_totp(user)

update(user, params)

Callback implementation for Pow.Context.update/2.

update_last_login_at(user)

update_user(user, attrs)

update_user_profile(user, attrs)

update_user_profile_sso_enabled(user, attrs)

validate_totp(user, totp_code)