TrentoWeb.Router.Helpers (Trento Web v2.3.1)

Module with named helpers generated from TrentoWeb.Router.

Functions

ability_path(conn_or_endpoint, action)

ability_path(conn_or_endpoint, action, params)

ability_url(conn_or_endpoint, action)

ability_url(conn_or_endpoint, action, params)

about_path(conn_or_endpoint, action)

about_path(conn_or_endpoint, action, params)

about_url(conn_or_endpoint, action)

about_url(conn_or_endpoint, action, params)

activity_log_path(conn_or_endpoint, action)

activity_log_path(conn_or_endpoint, action, params)

activity_log_url(conn_or_endpoint, action)

activity_log_url(conn_or_endpoint, action, params)

api_redirector_path(conn_or_endpoint, action, path)

api_redirector_path(conn_or_endpoint, action, path, params)

api_redirector_url(conn_or_endpoint, action, path)

api_redirector_url(conn_or_endpoint, action, path, params)

chart_path(conn_or_endpoint, action, id)

chart_path(conn_or_endpoint, action, id, params)

chart_url(conn_or_endpoint, action, id)

chart_url(conn_or_endpoint, action, id, params)

cluster_path(conn_or_endpoint, action)

cluster_path(conn_or_endpoint, action, params)

cluster_path(conn_or_endpoint, action, cluster_id, params)

cluster_url(conn_or_endpoint, action)

cluster_url(conn_or_endpoint, action, params)

cluster_url(conn_or_endpoint, action, cluster_id, params)

clusters_tagging_path(conn_or_endpoint, action, id)

clusters_tagging_path(conn_or_endpoint, action, id, params)

clusters_tagging_path(conn_or_endpoint, action, id, value, params)

clusters_tagging_url(conn_or_endpoint, action, id)

clusters_tagging_url(conn_or_endpoint, action, id, params)

clusters_tagging_url(conn_or_endpoint, action, id, value, params)

database_path(conn_or_endpoint, action)

database_path(conn_or_endpoint, action, params)

database_path(conn_or_endpoint, action, id, host_id, instance_number)

database_path(conn_or_endpoint, action, id, host_id, instance_number, params)

database_url(conn_or_endpoint, action)

database_url(conn_or_endpoint, action, params)

database_url(conn_or_endpoint, action, id, host_id, instance_number)

database_url(conn_or_endpoint, action, id, host_id, instance_number, params)

databases_tagging_path(conn_or_endpoint, action, id)

databases_tagging_path(conn_or_endpoint, action, id, params)

databases_tagging_path(conn_or_endpoint, action, id, value, params)

databases_tagging_url(conn_or_endpoint, action, id)

databases_tagging_url(conn_or_endpoint, action, id, params)

databases_tagging_url(conn_or_endpoint, action, id, value, params)

discovery_path(conn_or_endpoint, action)

discovery_path(conn_or_endpoint, action, params)

discovery_url(conn_or_endpoint, action)

discovery_url(conn_or_endpoint, action, params)

health_overview_path(conn_or_endpoint, action)

health_overview_path(conn_or_endpoint, action, params)

health_overview_url(conn_or_endpoint, action)

health_overview_url(conn_or_endpoint, action, params)

health_path(conn_or_endpoint, action)

health_path(conn_or_endpoint, action, params)

health_url(conn_or_endpoint, action)

health_url(conn_or_endpoint, action, params)

host_path(conn_or_endpoint, action)

host_path(conn_or_endpoint, action, params)

host_path(conn_or_endpoint, action, id, params)

host_url(conn_or_endpoint, action)

host_url(conn_or_endpoint, action, params)

host_url(conn_or_endpoint, action, id, params)

hosts_tagging_path(conn_or_endpoint, action, id)

hosts_tagging_path(conn_or_endpoint, action, id, params)

hosts_tagging_path(conn_or_endpoint, action, id, value, params)

hosts_tagging_url(conn_or_endpoint, action, id)

hosts_tagging_url(conn_or_endpoint, action, id, params)

hosts_tagging_url(conn_or_endpoint, action, id, value, params)

installation_path(conn_or_endpoint, action)

installation_path(conn_or_endpoint, action, params)

installation_url(conn_or_endpoint, action)

installation_url(conn_or_endpoint, action, params)

live_dashboard_path(conn_or_endpoint, action)

live_dashboard_path(conn_or_endpoint, action, params)

live_dashboard_path(conn_or_endpoint, action, page, params)

live_dashboard_path(conn_or_endpoint, action, node, page, params)

live_dashboard_url(conn_or_endpoint, action)

live_dashboard_url(conn_or_endpoint, action, params)

live_dashboard_url(conn_or_endpoint, action, page, params)

live_dashboard_url(conn_or_endpoint, action, node, page, params)

login_path(conn_or_endpoint, action)

login_path(conn_or_endpoint, action, params)

login_url(conn_or_endpoint, action)

login_url(conn_or_endpoint, action, params)

me_path(conn_or_endpoint, action)

me_path(conn_or_endpoint, action, params)

me_url(conn_or_endpoint, action)

me_url(conn_or_endpoint, action, params)

page_path(conn_or_endpoint, action, path)

page_path(conn_or_endpoint, action, path, params)

page_url(conn_or_endpoint, action, path)

page_url(conn_or_endpoint, action, path, params)

path(data, path)

Generates the path information including any necessary prefix.

profile_path(conn_or_endpoint, action)

profile_path(conn_or_endpoint, action, params)

profile_url(conn_or_endpoint, action)

profile_url(conn_or_endpoint, action, params)

prometheus_path(conn_or_endpoint, action)

prometheus_path(conn_or_endpoint, action, params)

prometheus_path(conn_or_endpoint, action, id, params)

prometheus_url(conn_or_endpoint, action)

prometheus_url(conn_or_endpoint, action, params)

prometheus_url(conn_or_endpoint, action, id, params)

refresh_path(conn_or_endpoint, action)

refresh_path(conn_or_endpoint, action, params)

refresh_url(conn_or_endpoint, action)

refresh_url(conn_or_endpoint, action, params)

render_spec_path(conn_or_endpoint, action)

render_spec_path(conn_or_endpoint, action, params)

render_spec_url(conn_or_endpoint, action)

render_spec_url(conn_or_endpoint, action, params)

sap_system_path(conn_or_endpoint, action)

sap_system_path(conn_or_endpoint, action, params)

sap_system_path(conn_or_endpoint, action, id, host_id, instance_number)

sap_system_path(conn_or_endpoint, action, id, host_id, instance_number, params)

sap_system_url(conn_or_endpoint, action)

sap_system_url(conn_or_endpoint, action, params)

sap_system_url(conn_or_endpoint, action, id, host_id, instance_number)

sap_system_url(conn_or_endpoint, action, id, host_id, instance_number, params)

sap_systems_tagging_path(conn_or_endpoint, action, id)

sap_systems_tagging_path(conn_or_endpoint, action, id, params)

sap_systems_tagging_path(conn_or_endpoint, action, id, value, params)

sap_systems_tagging_url(conn_or_endpoint, action, id)

sap_systems_tagging_url(conn_or_endpoint, action, id, params)

sap_systems_tagging_url(conn_or_endpoint, action, id, value, params)

settings_path(conn_or_endpoint, action)

settings_path(conn_or_endpoint, action, params)

settings_url(conn_or_endpoint, action)

settings_url(conn_or_endpoint, action, params)

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

suma_credentials_path(conn_or_endpoint, action)

suma_credentials_path(conn_or_endpoint, action, params)

suma_credentials_url(conn_or_endpoint, action)

suma_credentials_url(conn_or_endpoint, action, params)

suse_manager_path(conn_or_endpoint, action)

suse_manager_path(conn_or_endpoint, action, params)

suse_manager_path(conn_or_endpoint, action, advisory_name, params)

suse_manager_url(conn_or_endpoint, action)

suse_manager_url(conn_or_endpoint, action, params)

suse_manager_url(conn_or_endpoint, action, advisory_name, params)

swagger_ui_path(conn_or_endpoint, action)

swagger_ui_path(conn_or_endpoint, action, params)

swagger_ui_url(conn_or_endpoint, action)

swagger_ui_url(conn_or_endpoint, action, params)

url(data)

Generates the connection/endpoint base URL without any path information.

users_path(conn_or_endpoint, action)

users_path(conn_or_endpoint, action, params)

users_path(conn_or_endpoint, action, id, params)

users_url(conn_or_endpoint, action)

users_url(conn_or_endpoint, action, params)

users_url(conn_or_endpoint, action, id, params)