TestCommandedApp (Trento Web v1.2.0)

InMemory commanded app used in tests.

Functions

aggregate_state(aggregate_module, aggregate_uuid, timeout \\ 5000)

@spec aggregate_state(
  aggregate_module :: module(),
  aggregate_uuid :: Commanded.Aggregates.Aggregate.uuid(),
  timeout :: integer()
) :: Commanded.Aggregates.Aggregate.state()

Retrieve aggregate state of an aggregate.

Retrieving aggregate state is done by calling to the opened aggregate, or querying the event store for an optional state snapshot and then replaying the aggregate's event stream.

child_spec(opts)

@spec child_spec(opts :: Commanded.Application.options()) :: Supervisor.child_spec()

config()

Callback implementation for Commanded.Application.config/0.

start_link(opts \\ [])

Callback implementation for Commanded.Application.start_link/1.

stop(pid, timeout \\ 5000)

Callback implementation for Commanded.Application.stop/2.