Trento.Factory (Trento Web v1.2.0)

A simple Factory helper module to be used within tests to generate test data

Functions

application_instance_factory()

application_instance_registered_event_factory()

application_instance_without_host_factory()

build(factory_name, attrs \\ %{})

build_list(number_of_records, factory_name, attrs \\ %{})

build_pair(factory_name, attrs \\ %{})

check_result_factory()

checks_execution_requested_event_factory()

cluster_enrichment_data_factory()

cluster_factory()

cluster_registered_event_factory()

create(_)

@spec create(any()) :: no_return()

create(_, _)

@spec create(any(), any()) :: no_return()

create_list(_, _, _)

@spec create_list(any(), any(), any()) :: no_return()

create_pair(_, _)

@spec create_pair(any(), any()) :: no_return()

database_factory()

database_instance_factory()

database_instance_registered_event_factory()

database_instance_without_host_factory()

database_registered_event_factory()

discarded_discovery_event_factory()

discovery_event_factory()

factory(factory_name)

@spec factory(any()) :: no_return()

Raises a helpful error if no factory is defined.

hana_cluster_details_value_object()

heartbeat_factory()

host_added_to_cluster_event_factory()

host_checks_result_factory()

host_connection_settings_factory()

host_details_updated_event_factory()

host_factory()

host_registered_event_factory()

host_telemetry_factory()

insert(already_built_record)

insert(already_built_record, function_opts)

insert(factory_name, attrs, opts)

insert_list(number_of_records, factory_name, attrs \\ %{})

insert_list(number_of_records, factory_name, attrs, opts)

insert_pair(factory_name, attrs \\ %{})

insert_pair(factory_name, attrs, opts)

params_for(factory_name, attrs \\ %{})

params_with_assocs(factory_name, attrs \\ %{})

register_application_instance_command_factory()

register_cluster_host_factory()

register_database_instance_command_factory()

sap_system_factory()

sap_system_registered_event_factory()

sles_subscription_factory()

string_params_for(factory_name, attrs \\ %{})

string_params_with_assocs(factory_name, attrs \\ %{})

subscriptions_updated_event_factory()

tag_factory()