TrentoWeb.Router.Helpers (Trento Web v1.2.0)

Module with named helpers generated from TrentoWeb.Router.

Functions

about_path(conn_or_endpoint, action)

about_path(conn_or_endpoint, action, params)

about_url(conn_or_endpoint, action)

about_url(conn_or_endpoint, action, params)

catalog_path(conn_or_endpoint, action)

catalog_path(conn_or_endpoint, action, params)

catalog_url(conn_or_endpoint, action)

catalog_url(conn_or_endpoint, action, params)

cluster_path(conn_or_endpoint, action)

cluster_path(conn_or_endpoint, action, params)

cluster_path(conn_or_endpoint, action, cluster_id, params)

cluster_url(conn_or_endpoint, action)

cluster_url(conn_or_endpoint, action, params)

cluster_url(conn_or_endpoint, action, cluster_id, params)

clusters_tagging_path(conn_or_endpoint, action, id)

clusters_tagging_path(conn_or_endpoint, action, id, params)

clusters_tagging_path(conn_or_endpoint, action, id, value, params)

clusters_tagging_url(conn_or_endpoint, action, id)

clusters_tagging_url(conn_or_endpoint, action, id, params)

clusters_tagging_url(conn_or_endpoint, action, id, value, params)

databases_tagging_path(conn_or_endpoint, action, id)

databases_tagging_path(conn_or_endpoint, action, id, params)

databases_tagging_path(conn_or_endpoint, action, id, value, params)

databases_tagging_url(conn_or_endpoint, action, id)

databases_tagging_url(conn_or_endpoint, action, id, params)

databases_tagging_url(conn_or_endpoint, action, id, value, params)

discovery_path(conn_or_endpoint, action)

discovery_path(conn_or_endpoint, action, params)

discovery_url(conn_or_endpoint, action)

discovery_url(conn_or_endpoint, action, params)

health_overview_path(conn_or_endpoint, action)

health_overview_path(conn_or_endpoint, action, params)

health_overview_url(conn_or_endpoint, action)

health_overview_url(conn_or_endpoint, action, params)

host_path(conn_or_endpoint, action)

host_path(conn_or_endpoint, action, params)

host_path(conn_or_endpoint, action, id, params)

host_url(conn_or_endpoint, action)

host_url(conn_or_endpoint, action, params)

host_url(conn_or_endpoint, action, id, params)

hosts_tagging_path(conn_or_endpoint, action, id)

hosts_tagging_path(conn_or_endpoint, action, id, params)

hosts_tagging_path(conn_or_endpoint, action, id, value, params)

hosts_tagging_url(conn_or_endpoint, action, id)

hosts_tagging_url(conn_or_endpoint, action, id, params)

hosts_tagging_url(conn_or_endpoint, action, id, value, params)

installation_path(conn_or_endpoint, action)

installation_path(conn_or_endpoint, action, params)

installation_url(conn_or_endpoint, action)

installation_url(conn_or_endpoint, action, params)

live_dashboard_path(conn_or_endpoint, action)

live_dashboard_path(conn_or_endpoint, action, params)

live_dashboard_path(conn_or_endpoint, action, page, params)

live_dashboard_path(conn_or_endpoint, action, node, page, params)

live_dashboard_url(conn_or_endpoint, action)

live_dashboard_url(conn_or_endpoint, action, params)

live_dashboard_url(conn_or_endpoint, action, page, params)

live_dashboard_url(conn_or_endpoint, action, node, page, params)

page_path(conn_or_endpoint, action, path)

page_path(conn_or_endpoint, action, path, params)

page_url(conn_or_endpoint, action, path)

page_url(conn_or_endpoint, action, path, params)

path(data, path)

Generates the path information including any necessary prefix.

pow_session_path(conn_or_endpoint, action)

pow_session_path(conn_or_endpoint, action, params)

pow_session_url(conn_or_endpoint, action)

pow_session_url(conn_or_endpoint, action, params)

prometheus_path(conn_or_endpoint, action)

prometheus_path(conn_or_endpoint, action, params)

prometheus_path(conn_or_endpoint, action, id, params)

prometheus_url(conn_or_endpoint, action)

prometheus_url(conn_or_endpoint, action, params)

prometheus_url(conn_or_endpoint, action, id, params)

render_spec_path(conn_or_endpoint, action)

render_spec_path(conn_or_endpoint, action, params)

render_spec_url(conn_or_endpoint, action)

render_spec_url(conn_or_endpoint, action, params)

sap_system_path(conn_or_endpoint, action)

sap_system_path(conn_or_endpoint, action, params)

sap_system_url(conn_or_endpoint, action)

sap_system_url(conn_or_endpoint, action, params)

sap_systems_tagging_path(conn_or_endpoint, action, id)

sap_systems_tagging_path(conn_or_endpoint, action, id, params)

sap_systems_tagging_path(conn_or_endpoint, action, id, value, params)

sap_systems_tagging_url(conn_or_endpoint, action, id)

sap_systems_tagging_url(conn_or_endpoint, action, id, params)

sap_systems_tagging_url(conn_or_endpoint, action, id, value, params)

settings_path(conn_or_endpoint, action)

settings_path(conn_or_endpoint, action, params)

settings_url(conn_or_endpoint, action)

settings_url(conn_or_endpoint, action, params)

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

swagger_ui_path(conn_or_endpoint, action)

swagger_ui_path(conn_or_endpoint, action, params)

swagger_ui_url(conn_or_endpoint, action)

swagger_ui_url(conn_or_endpoint, action, params)

url(data)

Generates the connection/endpoint base URL without any path information.